moving components needed for details view
