+++ assets/src/dashboard/app/views/exploreTemplates/templateFilters.js
