+++ assets/src/dashboard/app/views/exploreTemplates/test/templateFilters.js
