+++ assets/src/dashboard/app/views/templateDetails/components.js
